Fix restart button on touch devices
When migrating the code for detecting swipes to hammer.js, I also
removed the constants that stored the previously event names. However,
eventTouchend was still used for buttons such that buttons would not
work on touch devices anymore. This fix migrates the button tap code to
hammer.js as well.