Work around missing tile labels on touch devices
The tile labels are displayed via the :hover pseudo class. This class is
never applied for touch input (it might still work for the first touch
point which is often used for mouse emulation). This workaround adds the
new CSS class '.hover' for tiles that is added on touchstart and removed
on touchend respectively touchcancel.